+++ composeApp/src/commonMain/kotlin/com/virtualworld/multiplatformiot/id/AppModule.kt
@@ -1,11 +1,13 @@
 package com.virtualworld.multiplatformiot.id
 
+import com.virtualworld.conectioninternet.data.conectionInternet.dataConectionInternetModule
 import com.virtualworld.multiplatformiot.feature.conectionInternet.id.featureConectionInternetModule
 import com.virtualworld.multiplatformiot.feature.menu.id.featureMenuModule
 import com.wirtualworld.multiplatformiot.feature.conectionLocal.featureConectionLocalModule
 import org.koin.dsl.module
 
 val appModule = module {
-    includes(featureMenuModule, featureConectionLocalModule,featureConectionInternetModule)
+    includes(featureMenuModule, featureConectionLocalModule,featureConectionInternetModule,
+        dataConectionInternetModule)
 
 }
