+++ data/conectionInternet/src/commonMain/kotlin/com/virtualworld/conectioninternet/data/conectionInternet/Module.kt
@@ -1,6 +1,7 @@
 package com.virtualworld.conectioninternet.data.conectionInternet
 
 import com.virtualworld.multiplatformiot.dataCoreModule
+import com.virtualworld.multiplatformiot.domain.conectionInternet.repository.RepocitoryInternet
 import dev.gitlive.firebase.Firebase
 import dev.gitlive.firebase.firestore.firestore
 import org.koin.core.module.dsl.singleOf
