+++ data/conectionInternet/src/commonMain/kotlin/com/virtualworld/conectioninternet/data/conectionInternet/RemoteDataSource.kt
@@ -1,6 +1,8 @@
 package com.virtualworld.conectioninternet.data.conectionInternet
 
 import com.virtualworld.multiplatformiot.ProductEmptyException
+import com.virtualworld.multiplatformiot.data.core.Arduino
+import com.virtualworld.multiplatformiot.data.core.NetworkResponseState
 import dev.gitlive.firebase.firestore.FirebaseFirestore
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
