--- data/conectionInternet/src/commonMain/kotlin/com/virtualworld/conectioninternet/data/conectionInternet/RepocitoryInternet.kt
@@ -1,11 +1,0 @@
-package com.virtualworld.conectioninternet.data.conectionInternet
-
-import kotlinx.coroutines.flow.Flow
-
-interface RepocitoryInternet {
-
-    fun getAllArduinos(usuario:String): Flow<NetworkResponseState<List<Arduino>>>
-
-
-
-}
