+++ domain/conectionInternet/src/commonMain/kotlin/com/virtualworld/multiplatformiot/domain/conectionInternet/repository/RepocitoryInternet.kt
@@ -1,0 +1,13 @@
+package com.virtualworld.multiplatformiot.domain.conectionInternet.repository
+
+import com.virtualworld.multiplatformiot.data.core.Arduino
+import com.virtualworld.multiplatformiot.data.core.NetworkResponseState
+import kotlinx.coroutines.flow.Flow
+
+interface RepocitoryInternet {
+
+    fun getAllArduinos(usuario:String): Flow<NetworkResponseState<List<Arduino>>>
+
+
+
+}
