+++ domain/conectionLocal/build.gradle.kts
@@ -1,35 +1,19 @@
 
-
 plugins {
     alias(libs.plugins.kotlinMultiplatform)
-
-
 }
 
 kotlin {
 
-
-
-    listOf(
-        iosX64(),
-        iosArm64(),
-        iosSimulatorArm64()
-    ).forEach { iosTarget ->
-        iosTarget.binaries.framework {
-            baseName = "DomainConectionLocal"
-            isStatic = true
-        }
-    }
-
-    jvm("desktop")
-
-
+    jvm()
 
     sourceSets {
 
         commonMain.dependencies {
 
-            implementation(projects.data.conectionLocal)
+            implementation(projects.data.core)
+
+            implementation(libs.kotlinx.coroutines.core)
 
             implementation(project.dependencies.platform(libs.koin.bom))
             implementation(libs.koin.core)
