--- domain/conectionLocal/src/androidMain/kotlin/com/virtualworld/multiplatformiot/Platform.android.kt
@@ -1,3 +1,0 @@
-package com.virtualworld.multiplatformiot
-
-actual fun platform() = "Android"
