+++ feature/conectionInternet/src/commonMain/kotlin/com/virtualworld/multiplatformiot/feature/conectionInternet/id/Module.kt
@@ -1,7 +1,6 @@
 package com.virtualworld.multiplatformiot.feature.conectionInternet.id
 
-import com.virtualworld.multiplatformiot.domain.conectionInternet.UseCaseInternet
-import com.virtualworld.multiplatformiot.domain.conectionInternet.domainConectionInternetModule
+import com.virtualworld.multiplatformiot.domain.conectionInternet.usecase.UseCaseInternet
 import com.virtualworld.multiplatformiot.feature.conectionInternet.screen.ConectionInternetViewModel
 import org.koin.core.module.dsl.factoryOf
 import org.koin.core.module.dsl.viewModelOf
@@ -10,11 +9,10 @@ import org.koin.dsl.module
 
 val featureConectionInternetModule = module {
 
-    includes( domainConectionInternetModule )
+
+    factoryOf(::UseCaseInternet)
 
     viewModelOf(::ConectionInternetViewModel)
-
-
 
 
 }
